Prometheus: Streaming exemplar sampling
