+++ pkg/tsdb/prometheus/querydata/framing_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/grafana/grafana/pkg/tsdb/prometheus/models"
 )
 
-var update = true
+var update = false
 
 func TestRangeResponses(t *testing.T) {
 	tt := []struct {
